refactor:#70 룸데이터 불러오는 부분 훅으로 리펙터링
